requirements.txt
Django==1.11.5
djangorestframework==3.6.4
Markdown==2.6.9
django-filter==1.0.4
django-extensions==1.9.0
coreapi==2.3.1
Pygments==2.2.0
Markdown==2.6.9
ipaddress==1.0.18
python-telegram-bot==7.0.1
pandas==0.20.3
